Added is_enabled() to LspTextCommand base class.
