+++ plugin/code_actions.py
@@ -8,7 +8,7 @@ except ImportError:
 
 
 from .core.clients import client_for_view
-from .core.configurations import is_supported_view, LspTextCommand
+from .core.clients import LspTextCommand
 from .core.protocol import Request, Range
 from .core.documents import get_position
 from .core.diagnostics import get_point_diagnostics
@@ -16,12 +16,8 @@ from .core.url import filename_to_uri
 
 
 class LspCodeActionsCommand(LspTextCommand):
-    def is_enabled(self, event=None):
-        if is_supported_view(self.view):
-            client = client_for_view(self.view)
-            if client and client.has_capability('codeActionProvider'):
-                return True
-        return False
+    def __init__(self, view):
+        super(LspCodeActionsCommand, self).__init__(view, 'codeActionProvider')
 
     def run(self, edit, event=None):
         client = client_for_view(self.view)
