+++ plugin/definition.py
@@ -1,21 +1,16 @@
 import sublime
 
-from .core.configurations import is_supported_view, LspTextCommand
+from .core.clients import LspTextCommand
 from .core.clients import client_for_view
 from .core.protocol import Request, Point
-from .core.documents import get_document_position, get_position, is_at_word
+from .core.documents import get_document_position, get_position
 from .core.url import uri_to_filename
 from .core.logging import debug
 
 
 class LspSymbolDefinitionCommand(LspTextCommand):
-    def is_enabled(self, event=None):
-        # TODO: check what kind of scope we're in.
-        if is_supported_view(self.view):
-            client = client_for_view(self.view)
-            if client and client.has_capability('definitionProvider'):
-                return is_at_word(self.view, event)
-        return False
+    def __init__(self, view):
+        super(LspSymbolDefinitionCommand, self).__init__(view, 'definitionProvider')
 
     def run(self, edit, event=None):
         client = client_for_view(self.view)
