+++ plugin/symbols.py
@@ -1,6 +1,6 @@
 
 from .core.protocol import SymbolKind
-from .core.configurations import is_supported_view, LspTextCommand
+from .core.clients import LspTextCommand
 from .core.clients import client_for_view
 from .core.protocol import Request, Range
 from .core.url import filename_to_uri
@@ -37,12 +37,8 @@ def format_symbol(item):
 
 
 class LspDocumentSymbolsCommand(LspTextCommand):
-    def is_enabled(self):
-        if is_supported_view(self.view):
-            client = client_for_view(self.view)
-            if client and client.has_capability('documentSymbolProvider'):
-                return True
-        return False
+    def __init__(self, view):
+        super(LspDocumentSymbolsCommand, self).__init__(view, 'documentSymbolProvider')
 
     def run(self, edit):
         client = client_for_view(self.view)
